feat(@keydown, method): create methods to increase/decrease user votes by pressing the up/down keys
Error:
- This way of doing it does not work, it is not reactive.
The code does not work because by passing `user.votes` directly to the `incrementVotes` and `decrementVotes` functions, you are modifying a copy of the votes value, not the reactive value inside the `user` object of the Vue state. To make it reactive, you must pass the entire object and update its `votes` property directly.

References:
- https://vueschool.notion.site/Exercise-3-Event-Listeners-a8acbc0476b443cfb41433f8b935eb0a